small typo

+++ DatabaseService/src/main/java/com/example/DatabaseService/repository/NotificationsRepository.java
@@ -3,4 +3,4 @@ package com.example.DatabaseService.repository;
 import com.example.DatabaseService.entity.Notifications;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface NotificationsRepository extends JpaRepository<Notifications, Integer> {}
+public interface NotificationsRepository extends JpaRepository<Notifications, Long> {}
